# openapi.yaml
openapi: 3.0.0
info:
  title: educultura API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3018/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /curso-cultural:
    get:
      tags:
        - CursoCultural
      summary: Listar todos los CursoCulturals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de CursoCulturals
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - CursoCultural
      summary: Crear CursoCultural
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: CursoCultural creado
  /curso-cultural/{id}:
    get:
      tags:
        - CursoCultural
      summary: Obtener CursoCultural por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: CursoCultural encontrado
        '404':
          description: CursoCultural no encontrado
    patch:
      tags:
        - CursoCultural
      summary: Actualizar CursoCultural
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: CursoCultural actualizado
    delete:
      tags:
        - CursoCultural
      summary: Eliminar CursoCultural
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: CursoCultural eliminado
  /curso-cultural/{id}/upload-image:
    post:
      tags:
        - CursoCultural
      summary: Subir imagen a CursoCultural
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...
  /estudiante:
    get:
      tags:
        - Estudiante
      summary: Listar todos los Estudiantes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Estudiantes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Estudiante
      summary: Crear Estudiante
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Estudiante creado
  /estudiante/{id}:
    get:
      tags:
        - Estudiante
      summary: Obtener Estudiante por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estudiante encontrado
        '404':
          description: Estudiante no encontrado
    patch:
      tags:
        - Estudiante
      summary: Actualizar Estudiante
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Estudiante actualizado
    delete:
      tags:
        - Estudiante
      summary: Eliminar Estudiante
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estudiante eliminado
  /estudiante/{id}/upload-image:
    post:
      tags:
        - Estudiante
      summary: Subir imagen a Estudiante
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...